fixed cart merge bug
